Improving error detection, resolution and communication.
This commit changes the default behavior of NULL aspects. It now always suspends the simulation and warns the user if this happens except when this is explicitly disabled (Normally it is best to first check if an aspect exists, but in some cases (highly repeated operations) it can be beneficial to mitigate the performance impact this has and instead first fetch the NULL aspect and handle this after.

Further error reporting and handling caused by aspect casting mismatches and errors in the protocol file on initiation have been improved. The split tree has been tweaked to perform better in high overlap scenarios.